% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.plot.R
\name{a.plot}
\alias{a.plot}
\title{a-plot DIF method}
\usage{
a.plot(a.R, a.F, purify, sig.level)
}
\arguments{
\item{a.R}{numeric: a-parameters for reference group}

\item{a.F}{numeric: a-parameters for focal group}

\item{purify}{logical: purification for the anchor}

\item{sig.level}{numeric: the significance level}
}
\value{
a-plot and a data frame including a_Reference
(a-parameters for reference group), a_Focal(a-parameters
for focal group), Distance(distance from the principal axis line),
Distance/SD (Distance divided by the standard deviation of the points
around the principal axis line), and Flagged(item(s) flagged as DIF).
}
\description{
a-plot DIF method
}
