% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.plot.R
\name{b.plot}
\alias{b.plot}
\title{The b-plot DIF procedure}
\usage{
b.plot(b.R, b.F, purify, sig.level)
}
\arguments{
\item{b.R}{numeric: b-parameters for reference group}

\item{b.F}{numeric: b-parameters for focal group}

\item{purify}{logical: purification for the anchor}

\item{sig.level}{numeric: the significance level}
}
\value{
b-plot and a data frame including b_Reference (b-parameters for reference group),
b_Focal(b-parameters for focal group), Distance(distance from the principal axis line),
Distance/SD (Distance devided by the standard deviation of the points around the principal axis line),
and Flagged(item(s) flagged as DIF).
}
\description{
The b-plot DIF procedure
}
