% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bring_output.R
\name{bring.flexmirt}
\alias{bring.flexmirt}
\alias{bring.bilog}
\alias{bring.parscale}
\alias{bring.mirt}
\title{Import Item and Ability Parameters from IRT Software}
\usage{
bring.flexmirt(file, type = c("par", "sco"), rePrm = TRUE, n.factor = 1)

bring.bilog(file, type = c("par", "sco"))

bring.parscale(file, type = c("par", "sco"))

bring.mirt(x)
}
\arguments{
\item{file}{A file name (including a directory) containing the item or ability parameters.}

\item{type}{A character string indicating a type of output file. Available types are "par" for a file
containing item parameter estimates and "sco" for a file containing ability parameter estimates.}

\item{rePrm}{A logical value. If TRUE, the item intercept and logit of item guessing parameters are
reparameterized into the item difficulty and item guessing parameters, respectively. To estimate the IRT
linking coefficients in this package, the item difficulty and item guessing parameters must be used.
Default is TRUE.}

\item{n.factor}{A numeric value indicating the number of estimated factors. This argument should be specified
when \code{type = "sco"}. Default is 1.}

\item{x}{An output object obtained from the function \code{\link[mirt]{mirt}}.}
}
\value{
These functions return a list including several objects. Only for the output of flexMIRT, the results of
multiple group analysis can be returned. In that case, each element of the list contains the estimation results for
each group.
}
\description{
These functions import item and/or ability parameters from BILOG-MG 3, PARSCALE 4, flexMIRT, and
mirt (R package).
}
\details{
The \code{\link{bring.flexmirt}} was written by modifying the function \code{\link[rpf]{read.flexmirt}}
(Pritikin, 2018). The functions \code{\link{bring.bilog}} and \code{\link{bring.parscale}}
were written by modifying the functions \code{\link[plink]{read.bilog}} and \code{read.parscale}
(Weeks, 2017), respectively.

The file extensions for item parameter and ability files, respectively, are: ".par" and ".sco"
for BILOG-MG and PARSCALE, and "-prm.txt" and "-sco.txt" for flexMIRT. For mirt, the name of the output
object is specified by the user.

Although \code{\link{bring.flexmirt}} is able to extract mutidimensional item and ability parameter estimates,
this package only deals with unidimensional IRT linking and equating methods.

For polytomous item parameters, \code{\link{bring.flexmirt}} and \code{\link{bring.mirt}} are able to import
the item parameters of the graded response model and the (generalized) partial credit model.
}
\note{
Regarding the item parameter files for any IRT software, only the internal object "full_df" in the returned list is
necessary for the IRT linking. The object "full_df" is a data.frame containing the item meta data
in a test form (e.g., item parameters, number of categories, models). See \code{\link{preplink}}, \code{\link{test.info}},
or \code{\link{simdat}} for more details about the item meta data.

Also, when item parameters are estimated using the partial credit or the generalized partial credit model,
item step parameters are returned in the object "full_df". Item step parameters are the overall item difficulty (or location)
parameter subtracted by the difficulty (or threshold) parameter for each category. See \code{\link{irtlink}} for more details
about the parameterization of the (generalized) partial credit model.
}
\section{Sample Output Files of IRT software}{


To illustrate how to import the item parameter estimate files of PARSCALE 4 and flexMIRT
using \code{\link{bring.parscale}} and \code{\link{bring.flexmirt}}, four item parameter
estimate output files are included in this package.

Among the four output files, two are from PARSCALE 4 with a file extension of ".PAR"
(i.e., "parscale_base.PAR" and "parscale_new.PAR") and the rest are from flexMIRT
with a file extension of "-prm.txt" (i.e., "flexmirt_base-prm.txt" and "flexmirt_new-prm.txt").

For the four IRT software, the base and new test forms are mixed-format tests with 55 items
consisting of fifty dichotomous items following the IRT 3PL model and five polytomous items with five
categories following the graded response model. In each test form, there are 15 common items
and they are items 41, 42, 42, ..., 54, 55. The examples below show how to import those output files.
}

\examples{
## example 1
## import the "-prm.txt" output file from flexMIRT
flex_base <- system.file("extdata", "flexmirt_base-prm.txt", package = "irteQ")
flex_new <- system.file("extdata", "flexmirt_new-prm.txt", package = "irteQ")
bring.flexmirt(file=flex_base, "par")$Group1$full_df
bring.flexmirt(file=flex_new, "par")$Group1$full_df

## example 2
## import the ".par" output file from PARSCALE
pscale_base <- system.file("extdata", "parscale_base.PAR", package = "irteQ")
pscale_new <- system.file("extdata", "parscale_new.PAR", package = "irteQ")
bring.parscale(file=pscale_base, "par")$full_df
bring.parscale(file=pscale_new, "par")$full_df

}
\references{
Cai, L. (2015). flexMIRT 3.0 Flexible multilevel multidimensional item analysis and test scoring [Computer software].
Chapel Hill, NC: Vector Psychometric Group.

Chalmers, R. P. (2012). mirt: A multidimensional item response theory package for the R environment.
\emph{Journal of Statistical Software, 48}(6), 1-29.

Weeks, J. P. (2010). plink: An R Package for Linking Mixed-Format Tests Using IRT-Based Methods.
\emph{Journal of Statistical Software, 35}(12), 1-33. URL http://www.jstatsoft.org/v35/i12/.

Pritikin, J. (2018). \emph{rpf: Response Probability Functions}. R package version 0.59.
https://CRAN.R-project.org/package=rpf

Muraki, E. & Bock, R. D. (2003). PARSCALE 4: IRT item analysis and test scoring for rating
scale data [Computer Program]. Chicago, IL: Scientific Software International. URL http://www.
ssicentral.com

Zimowski, M. F., Muraki, E., Mislevy, R. J., & Bock, R. D. (2003). BILOG-MG 3: Multiple-group
IRT analysis and test maintenance for binary items [Computer Program]. Chicago, IL: Scientific
Software International. URL http://www.ssicentral.com
}
\seealso{
\code{\link{irtlink}}
}
\author{
Hwanggyu Lim \email{hglim83@gmail.com}
}
