% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.for.D.R
\name{CI.for.D}
\alias{CI.for.D}
\title{The confidence interval for the MH delta effect size}
\usage{
CI.for.D(difMH.output, sig.level)
}
\arguments{
\item{difMH.output}{the data frame of the output from the statistics that are provided by the difMH function}

\item{sig.level}{numeric: the significance level}
}
\value{
The data frame of the output has a number of rows equal to the number of items.
The columns include DeltaMH (the transformed MH common odds ratio onto the Δ-metric),
SE(DeltaMH) (The standard error of deltaMH), Lower_Limit (the lower limits of the confidence interval for DeltaMH),
Upper_Limit (upper limits of the confidence interval for DeltaMH), Classification (Classification of DIF)
}
\description{
The confidence interval for the MH delta effect size
}
