% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.Logistic.MG.R
\name{DIF.Logistic.MG}
\alias{DIF.Logistic.MG}
\title{Logistic regression DIF method for more than two groups}
\usage{
DIF.Logistic.MG(data, grp, sig.level, purify, output.filename)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{grp}{factor:  the vector of group membership}

\item{sig.level}{numeric: the significance level}

\item{purify}{logical: Default is FALSE}

\item{output.filename}{character: either the file path or file name to save the output}
}
\description{
This code performs logistic regression testing for DIF using more than
two groups.
}
