% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF.Logistic.R
\name{DIF.Logistic}
\alias{DIF.Logistic}
\title{Logistic regression DIF method}
\usage{
DIF.Logistic(data, group, sig.level, purify, output.filename)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{group}{factor: the vector of group membership}

\item{sig.level}{numeric: the significance level}

\item{purify}{logical: Default is FALSE.}

\item{output.filename}{character: either the file path or file name to save the output}
}
\value{
Each list of the output has a data frame with the likelihood-ratio test statistic (G2),
its respective p-value, the change in Nagelkerke’s pseudo-R2 between two nested models.
In the data frame for both uniform and nonuniform DIF effect simultaneously, the DIF classification is reported.
}
\description{
Logistic regression DIF method
}
