% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMH.poly.R
\name{GMH.poly}
\alias{GMH.poly}
\title{The generlized MH procedure on polytomous data}
\usage{
GMH.poly(data, group, sig.level, purify)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{group}{factor: the vector of group membership}

\item{sig.level}{numeric: the significance level}

\item{purify}{logical: purification for the anchor}
}
\value{
The data frame of the output has a number of rows equla to the number of items.
The column include the GMH_Chi_Square (GMH Chi-square statistic), df (degree of freedom)
and p-value (p-value for the GMH Chi-square statistic).
}
\description{
The generlized MH procedure on polytomous data
}
