% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LordChi.R
\name{LordChi}
\alias{LordChi}
\title{Lord’s Chi-Square DIF Procedure}
\usage{
LordChi(
  par.Ref,
  par.Foc,
  cov.Ref,
  cov.Foc,
  fix.g = FALSE,
  same.scale = FALSE,
  purify = FALSE,
  sig.level = 0.05,
  delta.a = 0,
  delta.b = 0,
  delta.c = 0
)
}
\arguments{
\item{par.Ref}{a data frame for the reference group including the columns of item ID,
number of score categories, IRT model, a-parameter, b-parameter, and g-parameter}

\item{par.Foc}{a data frame for the focal group including the columns of item ID,
number of score categories, IRT model, a-parameter, b-parameter, and g-parameter}

\item{cov.Ref}{a list of covariance matrices for each item for the reference group}

\item{cov.Foc}{a list of covariance matrices for each item for the focal group}

\item{fix.g}{logical: whether g is fixed or constrained to be equal between the groups (Default is FALSE.)}

\item{same.scale}{logical: whether the item parameter estimates are on the same scale (Default is FALSE.)}

\item{purify}{logical: purification for the anchor (Default is FALSE.)}

\item{sig.level}{numeric: the significance level (Default is FALSE.)}

\item{delta.a}{(Default is 0.)}

\item{delta.b}{(Default is 0.)}

\item{delta.c}{(Default is 0.)}
}
\value{
A list that contains the metadata information for the reference group
($par.est.Ref), the metadata information($par.est.Focal.t), and test statistics
for the Lord’s chi-square($Lord.TS) and Wald tests ($Wald.TS) for the focal group
with the transformed item parameter estimates. The data frame for Lord.TS contains
for each item the item ID, number of score categories, and IRT model, as well as
  the Lord’s chi-square test statistic, degrees of freedom, its corresponding p-value,
  noncentrality parameter, and whether the item was flagged as DIF using the significance
  level supplied by the user.
}
\description{
Lord’s Chi-Square DIF Procedure
}
