% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mantel.poly.R
\name{Mantel.poly}
\alias{Mantel.poly}
\title{Mantel–Haenszel DIF Statistics for polytomous data}
\usage{
Mantel.poly(data, group, focal.name, ref.name, sig.level, purify)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{group}{factor: the vector of group membership}

\item{focal.name}{numeric or character: a single value}

\item{ref.name}{numeric or character: a single value}

\item{sig.level}{numeric: the significance level}

\item{purify}{logical: purification for the anchor}
}
\value{
The data frame of the output has a number of rows equla to the number of items.
The column include the Mantel_Chi_Square (Mantel Chi-square statistic)
and p-value (p-value for the Mantel Chi-square statistic).
}
\description{
Mantel–Haenszel DIF Statistics for polytomous data
}
