% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDIF.R
\name{PDIF}
\alias{PDIF}
\title{The PDIF effect size from the difMH}
\usage{
PDIF(data, group, focal.name, difMH.output, sig.level)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{group}{factor:  the vector of group membership}

\item{focal.name}{numeric or character: a single value}

\item{difMH.output}{character: either the file path or file name to save the output}

\item{sig.level}{numeric: the significance level}
}
\value{
The data frame of the output has a number of rows equal to the number of items.
The columns include Prop_Focal (proportion correct for the focal group),
Prop_Ref* (predicted proportion correct in the reference group), PDIF, SE(PDIF) (standard error of the PDIF),
Lower_Limit (lower limit of the confidence interval for the parameter),
and Upper_Limit (upper limit of the confidence interval for the parameter).
}
\description{
The PDIF effect size from the difMH
}
