% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtmodel.R
\name{plm}
\alias{plm}
\title{Polytomous Response Model Probabilities (GRM and GPCM)}
\usage{
plm(theta, a, d, D = 1, pmodel = c("GRM", "GPCM"))
}
\arguments{
\item{theta}{A vector of ability values.}

\item{a}{A numeric value of item discrimination (or slope) parameter.}

\item{d}{A vector of item threshold (or step) parameters.}

\item{D}{A scaling factor in IRT models to make the logistic function as close as possible to the normal ogive function  (if set to 1.7).
Default is 1.}

\item{pmodel}{A character string indicating the polytomous model being used. Available models are "GRM" for
the the graded response model and "GPCM" for the (generalized) partial credit model.}
}
\value{
This function returns a vector or matrix. When a matrix is returned, rows indicate theta values and columns represent
categories of an item.
}
\description{
This function computes the probability of selecting a specific category for an item
for a given set of theta values using the graded response model, partial credit model, and generalized
partial credit model.
}
\details{
When the category probabilities are computed for an item with the partial credit model, \code{a = 1} for that item.
When \code{pmodel = "GPCM"}, \code{d} should include step parameters. Item step parameters are the overall item
difficulty (or location) parameter subtracted by the difficulty (or threshold) parameter for each category. Thus, the number of step parameters
for an item with m categories is m-1 because a step parameter for the first category does not affect the category probabilities. For example,
if an item has five categories under the (generalized) partial credit model, four step parameters should be specified. For more details about
the parameterization of the (generalized) partial credit model, see \code{\link{irtlink}}.
}
\examples{
## Category probabilities for an item with four categories
## using a generalized partial credit model
plm(theta=c(-0.2, 0, 0.5), a=1.4, d=c(-0.2, 0, 0.5), D=1, pmodel='GPCM')

## Category probabilities for an item with five categories
## using a graded response model
plm(theta=c(-0.2, 0, 0.5), a=1.2, d=c(-0.4, -0.2, 0.4, 1.5), D=1, pmodel='GRM')

}
\seealso{
\code{\link{drm}}, \code{\link{irtlink}}
}
\author{
Hwanggyu Lim \email{hglim83@gmail.com}
}
