% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cond.exp.R
\name{plot.cond.exp}
\alias{plot.cond.exp}
\title{Plots of the conditional expected scores for polytomous item}
\usage{
plot.cond.exp(data, group, item, focal.name, ref.name, anchor)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{group}{factor:  the vector of group membership}

\item{item}{numeric or character: either the number or the name of the single item for which
the plots of conditional expected scores are plotted}

\item{focal.name}{numeric or character: a single value}

\item{ref.name}{numeric or character: a single value}

\item{anchor}{numeric: a vector of integer values specifying which items are currently considered
as anchor (DIF free) items.}
}
\value{
the plots of conditional expected scores for polytomous items in the reference group and in the focal group.
}
\description{
Plots of the conditional expected scores for polytomous item
}
