% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cond.p.R
\name{plot.cond.p}
\alias{plot.cond.p}
\title{The conditional p-value plot for the Standardized p-value difference (DSTD)}
\usage{
plot.cond.p(data, item, grp, focal.name, ref.name, anchor)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{item}{factor:  the vector of group membership}

\item{grp}{numeric or character: either the number or the name of the single item for which
the plots of conditional expected scores are plotted}

\item{focal.name}{numeric or character: a single value}

\item{ref.name}{numeric or character: a single value}

\item{anchor}{numeric: a vector of integer values specifying which items are currently considered
as anchor (DIF free) items.}
}
\value{
the plots of conditional p-value in the reference group and in the focal group
using the Standardized p-value difference (DSTD) method.
}
\description{
The conditional p-value plot for the Standardized p-value difference (DSTD)
}
