% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{IRT Linking to Transform Scale of Parameters}
\usage{
rescale(par.Ref, par.Foc, anchor.R, anchor.F)
}
\arguments{
\item{anchor.F}{}
}
\value{
A list of internal objects.
}
\description{
This function implements various IRT linking methods to transform the IRT scale of the new test form to the
IRT scale of the base test form based on unidimensional IRT models using Stocking-Lord method.
}
\author{
Hwanggyu Lim \email{hglim83@gmail.com}
}
