% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIBTEST.exp.R
\name{SIBTEST.exp}
\alias{SIBTEST.exp}
\title{SIBTEST}
\usage{
SIBTEST.exp(data, grp, focal.name, sig.level, purify)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{grp}{factor: the vector of group membership}

\item{focal.name}{numeric or character: a single value}

\item{sig.level}{numeric: the significance level}

\item{purify}{logical: purification for the anchor}
}
\value{
The output is returned as lists for uniform DIF and non-uniform DIF. Each list includes Item (item name),
focal_group (membership), n_matched_set (the matching items), n_suspect_set (the item to be tested for DIF),
beta (DIF measure) , SE (the standard error for beta), X2 (the Chi-square test statistics for DIF),
df (degree of freedom), p (statistical p-value).
}
\description{
SIBTEST
}
\references{
Chalmers, R. P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29. doi:10.18637/jss.v048.i06
Shealy, R. & Stout, W. (1993a). An item response theory model for test bias and differential item functioning. In P. Holland & H Wainer (Eds.), \emph{Differential item functioning} (pp. 197-240). Hillsdale, NJ: Erlbaum.
Shealy, R. & Stout, W. (1993b). A model-based standardization approach that separates true bias/DIF. \emph{Psychometrika, 58}, 159-194.
Zenisky, A. L., Hambleton, R. K., & Robin, F. (2003). Detection of differential item functioning in large-scale state assessments: A study evaluating a two-stage approach. \emph{Educational and Psychological Measurement, 63}, 51-64.
}
