% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMD.R
\name{SMD}
\alias{SMD}
\title{Standardized Mean Difference (SMD) for DIF detection}
\usage{
SMD(data, group, focal.name, ref.name, purify)
}
\arguments{
\item{data}{numeric: either the data matrix only, or the data matrix plus the vector of group membership.}

\item{group}{factor:  the vector of group membership}

\item{focal.name}{numeric or character: a single value}

\item{ref.name}{numeric or character: a single value}

\item{purify}{logical: Default is FALSE.}
}
\value{
The data frame of the output of SMD has a number of rows equal to the number of items.
The columns include the SMD,  SMD/Max_Cat, and  DIF_Classification.
SMD is the standardized mean difference estimate. SMD/Max_Cat is the SMD divided by maximum item score.
DIF_Classification has three effect sizes (“Negligible”, “Moderate”, or “Large”).
The effect size of an item is classified as “negligible” if SMD/Max_Cat < .05,
“moderate” .05 ≤ SMD/Max_Cat < .10, or “Large” SMD/Max_Cat  ≥ .10
}
\description{
Standardized Mean Difference (SMD) for DIF detection
}
