% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpar.R
\name{transpar}
\alias{transpar}
\title{Transform the IRT Parameters}
\usage{
transpar(x, constants)
}
\arguments{
\item{x}{A data.frame containing the item meta data (e.g., item parameters, number of categories, models ...). This data.frame
can be easily obtained using the function \code{\link{shape_df}}. See \code{\link{preplink}}, \code{\link{test.info}},
or \code{\link{simdat}} for more details about the item meta data.}

\item{constants}{A vector of length two containing the IRT linking coefficients (i.e., slope and intercept).
The first value is a slope and the second value is an intercept.}
}
\value{
This function returns a data.frame containing the transformed item parameters.
}
\description{
This function transforms item parameters using estimated IRT linking coefficients.
}
\examples{
## IRT liking using single-format tests
# use "KBneat" data set
dat_base <- KBneat$item$dat.base
dat_new <- KBneat$item$dat.new

# specify common items
common.b <- KBneat$common$common.b
common.n <- KBneat$common$common.b

# prepare a data set for IRT linking
x <- preplink(dat_base, dat_new, common.b, common.n)

# estimate IRT linking coefficients using the Mean/Sigma method
rst <- irtlink(x, method.link = 'SL', D = 1.7)

# extract the estimated linking coefficients
constants <- takelink(rst, "coefficients")

# transform the item parameters in the new test form
transpar(dat_new, constants)

}
\author{
Hwanggyu Lim \email{hglim83@gmail.com}
}
